+++ src/mocks/passthrough.handler.ts
@@ -1,0 +1,8 @@
+import { http } from 'msw'
+import { passthrough } from 'msw'
+
+export const passthroughHandlers = [
+  http.get('*.svg', () => passthrough()),
+  http.get('/src/components', () => passthrough()),
+  http.get('/src/assets/*', () => passthrough()),
+]
